feat(stylus): solver pipeline architectural modularization
