+++ EGoTouchService/Engine/StylusSolver/AsaTypes.hpp
