--- EGoTouchService/Engine/StylusSolver/CoorPostProcessor.cpp
@@ -1,319 +1,0 @@
-#include "CoorPostProcessor.h"
-#include <algorithm>
-#include <cmath>
-
-namespace Asa {
-
-void CoorPostProcessor::Reset() {
-    m_initialized = false;
-    m_frameCount = 0;
-    m_validHistory = 0;
-    m_history.fill(AsaCoorResult{});
-    m_prev[0] = m_prev[1] = AsaCoorResult{};
-    m_iirDim1Q8 = m_iirDim2Q8 = 0;
-    m_anchor = AsaCoorResult{};
-    m_offsetDim1 = m_offsetDim2 = 0;
-    m_lockDim1 = m_lockDim2 = false;
-    m_jitterActive = false;
-    m_speed = SpeedMetrics{};
-    m_lastIirCoefInt = 0;
-    m_motionFrameCount = 0;
-}
-
-// ══════════════════════════════════════════════
-// Step 1: PushHistory (TSACore: GetRealTimeCoor2Buf)
-// FIFO shift (index 0 = newest)
-// ══════════════════════════════════════════════
-void CoorPostProcessor::StepPushHistory(const AsaCoorResult& cur) {
-    for (int i = kHistoryLen - 1; i > 0; --i) {
-        m_history[static_cast<size_t>(i)] =
-            m_history[static_cast<size_t>(i - 1)];
-    }
-    m_history[0] = cur;
-    m_validHistory = std::min(m_validHistory + 1, kHistoryLen);
-    m_frameCount++;
-}
-
-// ══════════════════════════════════════════════
-// Step 2: 3-point average (TSACore: Get3PointAvgFilter)
-// ══════════════════════════════════════════════
-AsaCoorResult CoorPostProcessor::Step3PointAvg(
-        const AsaCoorResult& cur) {
-    if (!enable3PointAvg) return cur;
-    if (!m_prev[0].valid || !m_prev[1].valid) return cur;
-    AsaCoorResult out = cur;
-    out.dim1 = (m_prev[1].dim1 + m_prev[0].dim1 + cur.dim1) / 3;
-    out.dim2 = (m_prev[1].dim2 + m_prev[0].dim2 + cur.dim2) / 3;
-    return out;
-}
-
-// ══════════════════════════════════════════════
-// Step 3: Speed calculation (TSACore: GetCoorSpeed)
-// 24-frame path-accumulated speed with direction extraction
-// ══════════════════════════════════════════════
-void CoorPostProcessor::StepCalcSpeed() {
-    SpeedMetrics out{};
-    if (m_validHistory < 2) { m_speed = out; return; }
-
-    const int segments = m_validHistory - 1;
-    float accumPath = 0.0f;
-    // TSACore: accumulate total displacement per axis for directional detection
-    int32_t totalDisplaceDim1 = 0;
-    int32_t totalDisplaceDim2 = 0;
-
-    for (int i = 0; i < segments; ++i) {
-        const auto& a = m_history[static_cast<size_t>(i)];
-        const auto& b = m_history[static_cast<size_t>(i + 1)];
-        if (!a.valid || !b.valid) continue;
-
-        const float dx = static_cast<float>(a.dim1 - b.dim1);
-        const float dy = static_cast<float>(a.dim2 - b.dim2);
-        // TSACore: sqrt((dx² + dy²) * 100) then /10 for integer precision
-        // We approximate with float sqrt which is equivalent
-        const float segDist = std::sqrt(dx * dx + dy * dy);
-        accumPath += segDist;
-
-        if (i == 0) out.instant = segDist;
-        if (i == 2) out.shortAvg = accumPath / 3.0f;
-
-        totalDisplaceDim1 += (a.dim1 - b.dim1);
-        totalDisplaceDim2 += (a.dim2 - b.dim2);
-    }
-
-    out.fullAvg = accumPath / static_cast<float>(segments);
-    if (segments < 3) out.shortAvg = out.fullAvg;
-
-    // TSACore: DAT_1820dc30/dc34 = |total displacement| / frames
-    // Used by GetIIRCoef for directional override
-    if (segments > 0) {
-        out.avgVelDim1 = static_cast<float>(std::abs(totalDisplaceDim1)) /
-                         static_cast<float>(segments);
-        out.avgVelDim2 = static_cast<float>(std::abs(totalDisplaceDim2)) /
-                         static_cast<float>(segments);
-    }
-
-    // Motion state: update consecutive moving frame count
-    // Mirrors TSACore (DAT_18231950 bits 1+2): needs >= motionDetectFrames
-    // of continuous movement to enter "moving" mode
-    if (out.instant > 1.0f) {
-        m_motionFrameCount = std::min(m_motionFrameCount + 1, 100);
-    } else {
-        m_motionFrameCount = 0;
-    }
-
-    m_speed = out;
-}
-
-// ══════════════════════════════════════════════
-// Step 4: Dynamic IIR coefficient (TSACore: GetIIRCoef)
-// Uses still/moving state machine + directional override
-// Returns INTEGER coefficient (not normalized)
-// ══════════════════════════════════════════════
-int CoorPostProcessor::StepCalcIIRCoef(bool isInking) {
-    int lo, hi, lowThr;
-
-    // TSACore GetIIRCoef: (DAT_18231950 & 6) == 0 → Still mode, else Moving
-    // Still = hover/stationary: stronger smoothing (flash[0xa5e]/[0xa5f])
-    // Moving = inking/active:   weaker smoothing  (flash[0xa5c]/[0xa5d])
-    if (!isInking) {
-        // Still mode: stronger smoothing ([0xa5e]/[0xa5f])
-        lo = stillIirLow;
-        hi = stillIirHigh;
-        lowThr = stillLowSpeedThr;
-    } else {
-        // Moving mode: weaker smoothing ([0xa5c]/[0xa5d])
-        lo = movingIirLow;
-        hi = movingIirHigh;
-        lowThr = movingLowSpeedThr;
-    }
-
-    // TSACore: Directional velocity override
-    // When pen has consistent direction (DAT_18230a84/c34 != 0),
-    // override to still coefficients >> 1 (halved)
-    if (enableDirectionalHalve &&
-        (m_speed.avgVelDim1 > 1.0f || m_speed.avgVelDim2 > 1.0f)) {
-        hi = stillIirHigh >> 1;
-        lo = stillIirLow >> 1;
-    }
-
-    // TSACore: Integer speed-to-coefficient interpolation
-    // speed = DAT_1820dc38 (instant speed, already in integer scale)
-    const int speed = static_cast<int>(m_speed.instant);
-    int coef;
-
-    if (speed >= highSpeedThr) {
-        coef = hi;
-    } else if (speed < lowThr) {
-        coef = lo;
-    } else if (lowThr == highSpeedThr) {
-        // Edge case: thresholds equal → average
-        coef = (lo + hi) / 2;
-    } else {
-        // Linear interpolation (TSACore exact integer formula)
-        coef = lo + static_cast<int>(
-            (static_cast<int64_t>(hi - lo) * (speed - lowThr)) /
-            (highSpeedThr - lowThr));
-    }
-
-    m_lastIirCoefInt = coef;
-    return coef;
-}
-
-// ══════════════════════════════════════════════
-// Integer IIR core: (coef * cur + (N - coef) * prev) / N
-// Matches TSACore CoorIIRFilter exactly
-// ══════════════════════════════════════════════
-int32_t CoorPostProcessor::IIRFilterQ8(
-        int32_t prevQ8, int32_t curQ8, int coef, int N) {
-    // TSACore: return (coef * cur + (N - coef) * prev) / N
-    return (coef * curQ8 + (N - coef) * prevQ8) / N;
-}
-
-// ══════════════════════════════════════════════
-// Step 5: IIR filter (TSACore: CoorFilterProcess + CoorIIRFilterType)
-// Q8 fixed-point: coordinate × 256 + fractional remainder
-// Remainder is preserved across frames for sub-LSB precision
-// ══════════════════════════════════════════════
-AsaCoorResult CoorPostProcessor::StepIIR(
-        const AsaCoorResult& cur, int coefInt, bool skipIIR) {
-    // TSACore CoorFilterProcess: skip IIR when mode just transitioned
-    // (clears Q8 fractional remainder and passes through raw coordinate)
-    if (skipIIR) {
-        m_iirDim1Q8 = cur.dim1 << 8;
-        m_iirDim2Q8 = cur.dim2 << 8;
-        m_initialized = true;
-        return cur;
-    }
-
-    // TSACore: Skip IIR for first few frames (let coordinates settle)
-    // CoorFilterProcess: if ((DAT_18231c28 & 1)==0 || counters < 2) → zero fractional state
-    if (m_frameCount < iirSkipFrames) {
-        m_iirDim1Q8 = cur.dim1 << 8;
-        m_iirDim2Q8 = cur.dim2 << 8;
-        m_initialized = true;
-        return cur;
-    }
-
-    if (!m_initialized) {
-        m_iirDim1Q8 = cur.dim1 << 8;
-        m_iirDim2Q8 = cur.dim2 << 8;
-        m_initialized = true;
-        return cur;
-    }
-
-    const int N = std::max(1, iirDivisorN);
-
-    // TSACore CoorIIRFilterType:
-    //   prev_q8 = curFrac[idx] + prevCoor[idx] * 256   ← accumulated state
-    //   cur_q8  = curCoor[idx] << 8                     ← new input
-    //   result  = IIRFilter(prev_q8, cur_q8, coef, N)
-    //   output  = result >> 8                           ← integer coordinate
-    //   frac    = result & 0xFF                         ← remainder (kept)
-    //
-    // In our representation, m_iirDim1Q8 already holds
-    // the full Q8 state (integer part × 256 + fractional remainder).
-
-    m_iirDim1Q8 = IIRFilterQ8(m_iirDim1Q8, cur.dim1 << 8, coefInt, N);
-    m_iirDim2Q8 = IIRFilterQ8(m_iirDim2Q8, cur.dim2 << 8, coefInt, N);
-
-    AsaCoorResult out = cur;
-    out.dim1 = m_iirDim1Q8 >> 8;  // Integer part (truncate toward zero)
-    out.dim2 = m_iirDim2Q8 >> 8;
-    return out;
-}
-
-// ══════════════════════════════════════════════
-// Step 6: Jitter offset compensation (TSACore: AftCoorProcess)
-// Dynamic threshold based on sensor/screen dimensions
-// Independent X/Y axis locking
-// ══════════════════════════════════════════════
-AsaCoorResult CoorPostProcessor::StepJitter(
-        const AsaCoorResult& cur, bool isEdge) {
-    if (!enableJitter) return cur;
-
-    // TSACore AftCoorProcess: Edge detection in LOCAL space
-    // Edge = coordinate < 0x401 || coordinate >= (gridDim-1)*0x400
-    const bool isLocalEdge = isEdge ||
-        cur.dim1 < (kCoorUnit + 1) || cur.dim2 < (kCoorUnit + 1) ||
-        cur.dim1 >= (kGridDim - 1) * kCoorUnit ||
-        cur.dim2 >= (kGridDim - 1) * kCoorUnit;
-
-    // TSACore: Dynamic threshold = (param * gridDim * 0x400) / screenDim
-    // Uses DAT_1820d610 (gridDim=9) not the full sensor dimension
-    int32_t thrDim1, thrDim2;
-    if (isLocalEdge) {
-        thrDim1 = (screenDimDim1 > 0)
-            ? (jitterEdgeParamDim1 * kGridDim * kCoorUnit) / screenDimDim1
-            : 40;
-        thrDim2 = (screenDimDim2 > 0)
-            ? (jitterEdgeParamDim2 * kGridDim * kCoorUnit) / screenDimDim2
-            : 40;
-    } else {
-        thrDim1 = (screenDimDim1 > 0)
-            ? (jitterCenterParamDim1 * kGridDim * kCoorUnit) / screenDimDim1
-            : 20;
-        thrDim2 = (screenDimDim2 > 0)
-            ? (jitterCenterParamDim2 * kGridDim * kCoorUnit) / screenDimDim2
-            : 20;
-    }
-
-    // TSACore: Lock starts on first valid contact frame
-    // (DAT_18231b18 != 0 && DAT_18231c18 == 0)
-    if (!m_jitterActive) {
-        m_anchor = cur;
-        m_offsetDim1 = 0;
-        m_offsetDim2 = 0;
-        m_lockDim1 = true;
-        m_lockDim2 = true;
-        m_jitterActive = true;
-        return cur;
-    }
-
-    // TSACore: Independent X/Y axis lock check
-    // Each axis unlocks independently when displacement exceeds threshold
-    if (m_lockDim1) {
-        int32_t dx = cur.dim1 - m_anchor.dim1;
-        // TSACore: abs using (v ^ (v>>31)) - (v>>31)
-        if (std::abs(dx) > thrDim1) {
-            m_lockDim1 = false;
-        }
-        if (m_lockDim1) {
-            m_offsetDim1 = cur.dim1 - m_anchor.dim1;
-        }
-    }
-
-    if (m_lockDim2) {
-        int32_t dy = cur.dim2 - m_anchor.dim2;
-        if (std::abs(dy) > thrDim2) {
-            m_lockDim2 = false;
-        }
-        if (m_lockDim2) {
-            m_offsetDim2 = cur.dim2 - m_anchor.dim2;
-        }
-    }
-
-    // TSACore: output = coordinate - accumulated offset, clamped to [0, gridDim*0x400]
-    // Post-processing chain works in LOCAL space (range [0, 9*1024)),
-    // matching TSACore AftCoorProcess which clamps to DAT_1820d610 * 0x400.
-    AsaCoorResult out = cur;
-    int32_t resultDim1 = cur.dim1 - m_offsetDim1;
-    int32_t resultDim2 = cur.dim2 - m_offsetDim2;
-
-    // Clamp to LOCAL grid range (TSACore: DAT_1820d610/d611 * 0x400)
-    const int32_t maxDim1 = kGridDim * kCoorUnit;
-    const int32_t maxDim2 = kGridDim * kCoorUnit;
-    out.dim1 = std::clamp(resultDim1, 0, maxDim1);
-    out.dim2 = std::clamp(resultDim2, 0, maxDim2);
-    return out;
-}
-
-// ══════════════════════════════════════════════
-// Update 3-point history (after full chain)
-// ══════════════════════════════════════════════
-void CoorPostProcessor::StepUpdate3PtHistory(
-        const AsaCoorResult& result) {
-    m_prev[1] = m_prev[0];
-    m_prev[0] = result;
-}
-
-} // namespace Asa
