+++ EGoTouchService/Engine/StylusSolver/EdgeLiftCorrector.hpp
@@ -1,5 +1,6 @@
 #pragma once
-#include "AsaTypes.h"
+#include "AsaTypes.hpp"
+#include <cmath>
 #include <cstdint>
 
 namespace Asa {
@@ -33,17 +34,49 @@ public:
     /// @param sensorRows          Total sensor rows (for edge detection)
     /// @param sensorCols          Total sensor columns (for edge detection)
     /// @return true if this frame should be frozen (edge lift artifact)
-    bool IsEdgeLiftArtifact(
-        float curDim1, float curDim2,
-        float prevDim1, float prevDim2,
-        uint16_t curPressure, uint16_t prevPressure,
-        int sensorRows, int sensorCols) const;
+    inline bool IsEdgeLiftArtifact(
+            float curDim1, float curDim2,
+            float prevDim1, float prevDim2,
+            uint16_t curPressure, uint16_t prevPressure,
+            int sensorRows, int sensorCols) const {
+        if (!enabled) return false;
+
+        // Condition 1: pen just lifted (pressure dropped to 0)
+        if (curPressure != 0 || prevPressure == 0) return false;
+
+        // Condition 2: current coordinate is in the edge region
+        if (!IsInEdgeRegion(curDim1, curDim2, sensorRows, sensorCols))
+            return false;
+
+        // Condition 3: coordinate jump exceeds threshold
+        const float dx = curDim1 - prevDim1;
+        const float dy = curDim2 - prevDim2;
+        const float jumpDist = std::sqrt(dx * dx + dy * dy);
+        if (jumpDist < jumpThreshold) return false;
+
+        // Condition 4: previous frame was NOT in the edge region
+        // (the pen was in the interior and the coordinate "snapped" to the edge)
+        if (IsInEdgeRegion(prevDim1, prevDim2, sensorRows, sensorCols))
+            return false;
+
+        // All conditions met: this is an edge-lift artifact
+        return true;
+    }
 
     /// Check whether a coordinate is in the edge region.
     /// Used both internally and by StylusPipeline to set isEdge flag.
-    bool IsInEdgeRegion(
-        float dim1, float dim2,
-        int sensorRows, int sensorCols) const;
+    inline bool IsInEdgeRegion(
+            float dim1, float dim2,
+            int sensorRows, int sensorCols) const {
+        const float margin = edgeMarginCells * kCoorUnitF;
+        const float maxDim1 = static_cast<float>(sensorCols) * kCoorUnitF;
+        const float maxDim2 = static_cast<float>(sensorRows) * kCoorUnitF;
+
+        // Check if within margin cells of any border
+        if (dim1 < margin || dim1 > maxDim1 - margin) return true;
+        if (dim2 < margin || dim2 > maxDim2 - margin) return true;
+        return false;
+    }
 
     // ── Configuration ──
 
@@ -59,7 +92,7 @@ public:
     bool enabled = true;
 
 private:
-    static constexpr float kCoorUnit = static_cast<float>(Asa::kCoorUnit);
+    static constexpr float kCoorUnitF = static_cast<float>(Asa::kCoorUnit);
 };
 
 } // namespace Asa
